roles and permissions

+++ app/DataTables/AccreditationsDataTable.php
@@ -15,9 +15,16 @@ class AccreditationsDataTable extends DataTable
         return datatables()
         ->eloquent(app()->call([$this, 'query']))
         ->addColumn('action', function($item){
-            $action = '<a class="btn btn-success py-1 ps-2 pe-2" href="'.route('admin.accreditations.edit', [$item->id]).'" title="'.__('site.edit').'"><i class="fa fa-edit"></i></a>';
-            $action .= '<a class="btn btn-primary py-1 ps-2 pe-2 ms-1" href="'.route('admin.accreditations.show', [$item->id]).'" title="'.__('site.show').'"><i class="fa fa-eye"></i></a>';
-            $action .= '<button class="btn btn-danger py-1 ps-2 pe-2 ms-1" onclick="deleteItem(`'.route('admin.accreditations.destroy', [$item->id]).'`)" title="'.__('site.delete').'"><i class="fa fa-trash"></i></button>';
+            $action = '';
+            if (auth()->user()->can("credits_edit")){
+                $action .= '<a class="btn btn-success py-1 ps-2 pe-2" href="'.route('admin.accreditations.edit', [$item->id]).'" title="'.__('site.edit').'"><i class="fa fa-edit"></i></a>';
+            }
+            if (auth()->user()->can("credits_show")){
+                $action .= '<a class="btn btn-primary py-1 ps-2 pe-2 ms-1" href="'.route('admin.accreditations.show', [$item->id]).'" title="'.__('site.show').'"><i class="fa fa-eye"></i></a>';
+            }
+            if (auth()->user()->can("credits_delete")){
+                $action .= '<button class="btn btn-danger py-1 ps-2 pe-2 ms-1" onclick="deleteItem(`'.route('admin.accreditations.destroy', [$item->id]).'`)" title="'.__('site.delete').'"><i class="fa fa-trash"></i></button>';
+            }
             return $action;
         })
         ->editColumn('title', function($item){
@@ -63,6 +70,29 @@ class AccreditationsDataTable extends DataTable
                 "url"=> asset('lang/arabic.json')  
             ];
         }
+        $buttons= [
+            Button::make([
+                "extend"=> "export",
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-download\"></i> ".__('site.export')." &nbsp;<span class=\"caret\"/>'}"
+            ])->addClass("btn btn-danger"),
+
+            Button::make([
+                "extend"=> "print",
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-print\"></i> ".__('site.print')."'}"
+            ])->addClass("btn btn-info"),
+            Button::make([
+                "extend"=> 'reload',
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-recycle\"></i> ".__('site.reload')."'}"
+            ])->addClass("btn btn-success")
+        ];
+        if (auth()->user()->can("credits_create")){
+            $it = Button::make([
+                "extend"=> "create",
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-plus\"></i> ".__('site.create')."'}",
+                "init" => "function(api, node, config){ $(node).removeClass('btn-secondary'); }"
+            ])->addClass("btn btn-primary");
+            array_unshift($buttons, $it);
+        }
         return $this->builder()
                     ->setTableId('items-table')
                     ->columns($this->getColumns())
@@ -72,26 +102,7 @@ class AccreditationsDataTable extends DataTable
                     ->parameters([
                         "language" => $lang,
                     ])
-                    ->buttons(
-                        Button::make([
-                            "extend"=> "create",
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-plus\"></i> ".__('site.create')."'}",
-                            "init" => "function(api, node, config){ $(node).removeClass('btn-secondary'); }"
-                        ])->addClass("btn btn-primary"),
-                        Button::make([
-                            "extend"=> "export",
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-download\"></i> ".__('site.export')." &nbsp;<span class=\"caret\"/>'}"
-                        ])->addClass("btn btn-danger"),
-
-                        Button::make([
-                            "extend"=> "print",
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-print\"></i> ".__('site.print')."'}"
-                        ])->addClass("btn btn-info"),
-                        Button::make([
-                            "extend"=> 'reload',
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-recycle\"></i> ".__('site.reload')."'}"
-                        ])->addClass("btn btn-success")
-                    );
+                    ->buttons($buttons);
     }
 
     /**
