+++ app/DataTables/ExtrasDataTable.php
@@ -19,9 +19,12 @@ class ExtrasDataTable extends DataTable
             return $item->updated_at->format('Y-m-d');
         })
         ->addColumn('action', function($item){
-            $action = '<a class="btn btn-success py-1 ps-2 pe-2" href="'.route('admin.extra.edit', [$item->id]).'" title="'.__('site.edit').'"><i class="fa fa-edit"></i></a>';
-            $action .= '<a class="btn btn-primary py-1 ps-2 pe-2 ms-1" href="'.route('admin.extra.show', [$item->id]).'" title="'.__('site.show').'"><i class="fa fa-eye"></i></a>';
-            $action .= '<button class="btn btn-danger py-1 ps-2 pe-2 ms-1" onclick="deleteItem(`'.route('admin.extra.destroy', [$item->id]).'`)" title="'.__('site.delete').'"><i class="fa fa-trash"></i></button>';
+            $action = "";
+            if (auth()->user()->can("about_components")){
+                $action = '<a class="btn btn-success py-1 ps-2 pe-2" href="'.route('admin.extra.edit', [$item->id]).'" title="'.__('site.edit').'"><i class="fa fa-edit"></i></a>';
+                $action .= '<a class="btn btn-primary py-1 ps-2 pe-2 ms-1" href="'.route('admin.extra.show', [$item->id]).'" title="'.__('site.show').'"><i class="fa fa-eye"></i></a>';
+                $action .= '<button class="btn btn-danger py-1 ps-2 pe-2 ms-1" onclick="deleteItem(`'.route('admin.extra.destroy', [$item->id]).'`)" title="'.__('site.delete').'"><i class="fa fa-trash"></i></button>';
+            }
             return $action;
         })
         ->editColumn('title', function($item){
@@ -63,10 +66,31 @@ class ExtrasDataTable extends DataTable
                 "url"=> asset('lang/arabic.json')  
             ];
         }
+        $buttons= [
+            Button::make([
+                "extend"=> "export",
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-download\"></i> ".__('site.export')." &nbsp;<span class=\"caret\"/>'}"
+            ])->addClass("btn btn-danger"),
 
-
+            Button::make([
+                "extend"=> "print",
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-print\"></i> ".__('site.print')."'}"
+            ])->addClass("btn btn-info"),
+            Button::make([
+                "extend"=> 'reload',
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-recycle\"></i> ".__('site.reload')."'}"
+            ])->addClass("btn btn-success")
+        ];
+        if (auth()->user()->can("about_components")){
+            $it = Button::make([
+                "extend"=> "create",
+                "text"=> "function(dt, button, config){ return '<i class=\"fa fa-plus\"></i> ".__('site.create')."'}",
+                "init" => "function(api, node, config){ $(node).removeClass('btn-secondary'); }"
+            ])->addClass("btn btn-primary");
+            array_unshift($buttons, $it);
+        }
         return $this->builder()
-                    ->setTableId('extras-table')
+                    ->setTableId('items-table')
                     ->columns($this->getColumns())
                     ->minifiedAjax()
                     ->dom('Bfrtip')
@@ -74,26 +98,7 @@ class ExtrasDataTable extends DataTable
                     ->parameters([
                         "language" => $lang,
                     ])
-                    ->buttons(
-                        Button::make([
-                            "extend"=> "create",
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-plus\"></i> ".__('site.create')."'}",
-                            "init" => "function(api, node, config){ $(node).removeClass('btn-secondary'); }"
-                        ])->addClass("btn btn-primary"),
-                        Button::make([
-                            "extend"=> "export",
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-download\"></i> ".__('site.export')." &nbsp;<span class=\"caret\"/>'}"
-                        ])->addClass("btn btn-danger"),
-
-                        Button::make([
-                            "extend"=> "print",
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-print\"></i> ".__('site.print')."'}"
-                        ])->addClass("btn btn-info"),
-                        Button::make([
-                            "extend"=> 'reload',
-                            "text"=> "function(dt, button, config){ return '<i class=\"fa fa-recycle\"></i> ".__('site.reload')."'}"
-                        ])->addClass("btn btn-success")
-                    );
+                    ->buttons($buttons);
     }
 
     /**
